+++ services/instance.go
@@ -1,18 +1,18 @@
 package services
 
-import "github.com/namely/broadway/domain"
+import "github.com/namely/broadway/broadway"
 
 // InstanceService definition
 type InstanceService struct {
-	repo domain.InstanceRepository
+	repo broadway.InstanceRepository
 }
 
 // NewInstanceService creates a new instance service
-func NewInstanceService(r domain.InstanceRepository) *InstanceService {
+func NewInstanceService(r broadway.InstanceRepository) *InstanceService {
 	return &InstanceService{repo: r}
 }
 
 // Create a new instance
-func (is *InstanceService) Create(i domain.Instance) error {
+func (is *InstanceService) Create(i broadway.Instance) error {
 	return is.repo.Save(i)
 }
